+++ backend/libraries/chat_events/src/chat_events.rs
@@ -16,7 +16,7 @@ use serde::{Deserialize, Serialize};
 use sha2::{Digest, Sha256};
 use std::cmp::{max, Reverse};
 use std::collections::hash_map::Entry::{Occupied, Vacant};
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use types::{Hash, MessageReport, ReportedMessageInternal};
 
 pub const OPENCHAT_BOT_USER_ID: UserId = UserId::new(Principal::from_slice(&[228, 104, 142, 9, 133, 211, 135, 217, 129, 1]));
@@ -657,17 +657,14 @@ impl ChatEvents {
 
         if let Some((message, index)) = self.message_internal_mut(EventIndex::default(), None, message_id.into(), now) {
             if let MessageContentInternal::ReportedMessage(r) = &mut message.content {
-                if r.users.insert(user_id) {
-                    if r.reports.len() < 10 {
-                        r.reports.push(MessageReport {
-                            reported_by: user_id,
-                            timestamp: now,
-                            reason_code,
-                            notes,
-                        });
-                    }
-                    self.mark_event_updated(None, index, now);
-                }
+                r.reports.retain(|x| x.reported_by != user_id);
+                r.reports.push(MessageReport {
+                    reported_by: user_id,
+                    timestamp: now,
+                    reason_code,
+                    notes,
+                });
+                self.mark_event_updated(None, index, now);
                 return;
             }
         }
@@ -683,7 +680,6 @@ impl ChatEvents {
                     reason_code,
                     notes,
                 }],
-                users: HashSet::from([user_id]),
             }),
             replies_to: Some(ReplyContext {
                 event_list_if_other: Some((chat_id, thread_root_message_index)),